Added the authentication and starting to add error handling

+++ backend/routes/posts.js
@@ -44,6 +44,10 @@ router.post('', checkAuth, multer({storage}).single("image"), (req,res,next) => 
                 id: createdPost._id
             }
         });
+    }).catch(error => {
+        res.status(500).json({
+            message: 'Creating a post failed!'
+        })
     });
 });
 
@@ -57,11 +61,15 @@ router.put('/:id', checkAuth, multer({storage}).single("image"), (req,res,next) 
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    console.log(post);
-    Post.updateOne({_id: req.params.id},post).then(result => {
-        res.status(200).json({message: 'Updated Successful'});
+    Post.updateOne({_id: req.params.id, creator: req.userData.userId},post).then(result => {
+        if(result.n > 0){
+            res.status(200).json({message: 'Updated Successful'});
+        } else {
+            res.status(401).json({message: 'Not authorized!'});
+        }
     });
 });
 
@@ -100,9 +108,12 @@ router.get('',(req,res,next) => {
 });
 
 router.delete('/:id', checkAuth,(req,res,next) => {
-    Post.deleteOne({_id: req.params.id}).then(result => {
-        console.log(result);
-        res.status(200).json({ message: 'Post deleted' });
+    Post.deleteOne({_id: req.params.id, creator: req.userData.userId}).then(result => {
+        if(result.n > 0){
+            res.status(200).json({ message: 'Post deleted' });
+        } else {
+            res.status(401).json({message: 'Not authorized!'});
+        }
     });
 });
